Refactor ElasticSearchBackend to make it testable
